requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
click==8.1.8
ecdsa==0.19.0
fastapi==0.115.9
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
